Rename to seq2seq_base and updated hf watch to get config from get_datalogger()
