+++ dataquality/loggers/data_logger/seq2seq/seq2seq_base.py
@@ -13,7 +13,7 @@ from dataquality.loggers.data_logger.base_data_logger import (
     DataSet,
     MetasType,
 )
-from dataquality.loggers.logger_config.seq2seq.seq2seq import (
+from dataquality.loggers.logger_config.seq2seq.seq2seq_base import (
     Seq2SeqLoggerConfig,
     seq2seq_logger_config,
 )
