+++ dataquality/loggers/model_logger/__init__.py
@@ -6,7 +6,7 @@ from dataquality.loggers.model_logger import (
     text_ner,
 )
 from dataquality.loggers.model_logger.base_model_logger import BaseGalileoModelLogger
-from dataquality.loggers.model_logger.seq2seq import encoder_decoder, seq2seq
+from dataquality.loggers.model_logger.seq2seq import encoder_decoder, seq2seq_base
 
 __all__ = [
     "image_classification",
@@ -15,6 +15,6 @@ __all__ = [
     "text_multi_label",
     "text_ner",
     "BaseGalileoModelLogger",
-    "seq2seq",
+    "seq2seq_base",
     "encoder_decoder",
 ]
