+++ dataquality/loggers/model_logger/seq2seq/encoder_decoder.py
@@ -6,7 +6,7 @@ from dataquality.loggers.logger_config.seq2seq.encoder_decoder import (
     EncoderDecoderLoggerConfig,
     encoder_decoder_logger_config,
 )
-from dataquality.loggers.model_logger.seq2seq.seq2seq import Seq2SeqModelLogger
+from dataquality.loggers.model_logger.seq2seq.seq2seq_base import Seq2SeqModelLogger
 
 
 class EncoderDecoderModelLogger(Seq2SeqModelLogger):
