+++ tests/loggers/test_seq2seq.py
@@ -15,7 +15,7 @@ from dataquality.loggers.data_logger.base_data_logger import DataSet
 from dataquality.loggers.data_logger.seq2seq.encoder_decoder import (
     EncoderDecoderDataLogger,
 )
-from dataquality.loggers.data_logger.seq2seq.seq2seq import Seq2SeqDataLogger
+from dataquality.loggers.data_logger.seq2seq.seq2seq_base import Seq2SeqDataLogger
 from dataquality.loggers.logger_config.seq2seq.encoder_decoder import (
     encoder_decoder_logger_config,
 )
